TraceQL search: the benchmark behind the two derived tables' figures (#560)
docs/benchmarks/issue560-two-table-reads.sh builds corpus C1 with the
design record's corpus statements and corpus C2 from its own block, and
prints one key and value per figure: the storage of both tables and of
six codecs, the tail-of-bucket and ceiling measurements, the fragmented
table read before a merge, and the rows, bytes and marks the two
generators read against the span scan. It refuses to run from CI.

Also two lint fixes in the new tests.

+++ crates/pulsus-read/tests/traces_recency_table_literals.rs
@@ -115,8 +115,7 @@ fn the_error_view_filters_on_the_code_the_reader_renders_for_error() {
         panic!("trace_error_spans_mv in {CATALOG} carries no `WHERE status_code = <n>`:\n{def}")
     });
     assert_eq!(
-        n,
-        i64::from(m),
+        n, m,
         "trace_error_spans_mv in {CATALOG} keeps spans with status_code = {n}; the reader \
          renders status_code != {m} for `{{ status != error }}`"
     );
